refactor: remove obsolete KeyArea files and store
- Remove old KeyArea.tsx component
- Remove obsolete hooks: useKeyArea, useKeyInteraction, useKeyState, useLongPress, useRippleEffect
- Remove old KeyInfoDisplay and SVGRippleEffect components
- Remove obsolete constants and utilities: keyArea.ts, animations.ts
- Remove old circle-of-fifths store implementation
- Remove associated test files

These files have been replaced by the new unified KeyArea architecture
with improved separation of concerns and reduced complexity.
