+++ src/features/circle-of-fifths/hooks/useCircleOfFifths.ts
@@ -1,6 +1,6 @@
-import { useMemo } from 'react';
 import { CIRCLE_LAYOUT, TEXT_RADIUS } from '../constants';
-import { Point, SegmentPaths } from '@/features/circle-of-fifths/types';
+import type { Point } from '@/shared/types/graphics';
+import type { SegmentPaths } from '../types';
 import { calculateTextPosition, calculateTextRotation } from '../utils/geometry';
 import { generateThreeSegmentPaths } from '../utils/pathGeneration';
 import { CircleOfFifthsService, CircleSegmentDTO } from '@/domain/services/CircleOfFifths';
@@ -32,44 +32,44 @@ export interface SegmentData {
  * textRotation: number;
  * }} 描画に必要なプロパティを含むオブジェクト。
  */
+// 固定値の事前計算（モジュールレベルで一度だけ実行）
+const PRECOMPUTED_VIEW_BOX = (() => {
+  const size = CIRCLE_LAYOUT.RADIUS * 2;
+  return `-${CIRCLE_LAYOUT.RADIUS} -${CIRCLE_LAYOUT.RADIUS} ${size} ${size}`;
+})();
+
+const PRECOMPUTED_SEGMENTS: SegmentData[] = CircleOfFifthsService.getSegmentDTOs().map(segment => {
+  const { position } = segment;
+
+  // 3つの同心円弧からなるセグメントのパスを生成
+  const paths = generateThreeSegmentPaths(
+    position,
+    CIRCLE_LAYOUT.INNER_RADIUS,
+    CIRCLE_LAYOUT.MIDDLE_RADIUS,
+    CIRCLE_LAYOUT.RADIUS
+  );
+
+  // 各領域の中心にテキストを配置するための座標を計算
+  const textPositions = {
+    minorTextPos: calculateTextPosition(position, TEXT_RADIUS.MINOR),
+    majorTextPos: calculateTextPosition(position, TEXT_RADIUS.MAJOR),
+    signatureTextPos: calculateTextPosition(position, TEXT_RADIUS.SIGNATURE),
+  };
+
+  return {
+    segment,
+    paths,
+    textPositions,
+  };
+});
+
+const PRECOMPUTED_TEXT_ROTATION = calculateTextRotation();
+
 export const useCircleOfFifths = () => {
-  // SVGのビューボックス属性を計算。円がアートボードにちょうど収まるように設定。
-  const viewBox = useMemo(() => {
-    const size = CIRCLE_LAYOUT.RADIUS * 2;
-    return `-${CIRCLE_LAYOUT.RADIUS} -${CIRCLE_LAYOUT.RADIUS} ${size} ${size}`;
-  }, []); // CIRCLE_LAYOUTは不変な定数なので、依存配列は空で初回のみ計算。
-
-  // 全セグメントの描画データを一度の計算でまとめて生成し、メモ化する。(再レンダリング時の不要な計算コストを削減)
-  const segments = useMemo<SegmentData[]>(() => {
-    return CircleOfFifthsService.getSegmentDTOs().map(segment => {
-      const { position } = segment;
-
-      // 3つの同心円弧からなるセグメントのパスを生成
-      const paths = generateThreeSegmentPaths(
-        position,
-        CIRCLE_LAYOUT.INNER_RADIUS,
-        CIRCLE_LAYOUT.MIDDLE_RADIUS,
-        CIRCLE_LAYOUT.RADIUS
-      );
-
-      // 各領域の中心にテキストを配置するための座標を計算
-      const textPositions = {
-        minorTextPos: calculateTextPosition(position, TEXT_RADIUS.MINOR),
-        majorTextPos: calculateTextPosition(position, TEXT_RADIUS.MAJOR),
-        signatureTextPos: calculateTextPosition(position, TEXT_RADIUS.SIGNATURE),
-      };
-
-      return {
-        segment,
-        paths,
-        textPositions,
-      };
-    });
-  }, []);
-
-  // テキスト要素を円の向きに合わせるための統一の回転角度を計算。
-  const textRotation = useMemo(() => calculateTextRotation(), []);
-
-  // 算出された全ての値をオブジェクトとして集約し、コンポーネントに提供する。
-  return { viewBox, segments, textRotation };
+  // 事前計算済みの固定値をそのまま返す（メモ化不要）
+  return {
+    viewBox: PRECOMPUTED_VIEW_BOX,
+    segments: PRECOMPUTED_SEGMENTS,
+    textRotation: PRECOMPUTED_TEXT_ROTATION,
+  };
 };
