+++ src/features/circle-of-fifths/utils/index.ts
@@ -9,5 +9,4 @@
 export * as validation from './validation';
 export * as geometry from './geometry';
 export * as pathGeneration from './pathGeneration';
-export * as animations from './animations';
 export * as classNames from './classNames';
