+++ src/features/circle-of-fifths/utils/pathGeneration.ts
@@ -1,4 +1,5 @@
-import { SegmentPaths, CircleOfFifthsError } from '@/features/circle-of-fifths/types';
+import type { SegmentPaths } from '../types';
+import { CircleOfFifthsError } from '@/features/circle-of-fifths/types';
 import { CIRCLE_LAYOUT } from '../constants/index';
 import { isValidPosition } from './validation';
 import { calculateAngle, normalizeAngle, polarToCartesian } from './geometry';
